pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "airtable-databricks-connector"
version = "1.0.0"
description = "Production-ready Airtable to Databricks Delta Lake connector following Lakeflow Community Connectors standard"
authors = [{name = "Kaustav Paul", email = "kaustav.paul@databricks.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["databricks", "airtable", "delta-lake", "lakeflow", "connector", "etl"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

[project.urls]
Homepage = "https://github.com/databrickslabs/lakeflow-community-connectors"
Documentation = "https://github.com/databrickslabs/lakeflow-community-connectors/blob/master/sources/airtable/README.md"
Repository = "https://github.com/databrickslabs/lakeflow-community-connectors"
Issues = "https://github.com/databrickslabs/lakeflow-community-connectors/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "integration: marks tests as integration tests (require live Airtable connection)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | node_modules
  | ui
  | dist
  | build
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # Line too long (handled by black)
    "B008",  # Do not perform function calls in argument defaults
]
exclude = [
    ".git",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    "node_modules",
    "ui",
    "dist",
    "build",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Imported but unused (module imports)
"tests/*" = ["S101"]      # Use of assert detected

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_optional = true
exclude = [
    "ui/",
    "tests/",
    "node_modules/",
    ".venv/",
]

[[tool.mypy.overrides]]
module = [
    "pyspark.*",
    "delta.*",
    "dlt.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["sources", "pipeline", "libs"]
omit = [
    "tests/*",
    "ui/*",
    "*/__pycache__/*",
    "*/node_modules/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

